Modal de confirmación al eliminar un grupo.
Ademas cambio el metodo para eliminar un grupo.

+++ app/controllers/Grupos.java
@@ -547,32 +547,42 @@ public class Grupos extends Controller {
 	 *
 	 * @return
 	 */
-	public static Result eliminaGrupo(Long id) {
-		// Obtiene el directorio del grupo a eliminar
-		File file = new File("./public/grupos/" + id.toString());
-		// Obtiene todas las reuniones del grupo.
-		List<Reunion> reuniones = Reunion.getReuniones(id);
+	public static Result eliminaGrupo() {
+		if (!verificaSession()) {
+			return redirect(routes.Application.index());
+		} else {
+			Form<Grupo> formGrupo = form(Grupo.class).bindFromRequest();
+			if (formGrupo.hasErrors()) {
+				return badRequest();
+			} else {
+				Long id = formGrupo.get().id;
+				// Obtiene el directorio del grupo a eliminar
+				File file = new File("./public/grupos/" + id.toString());
+				// Obtiene todas las reuniones del grupo.
+				List<Reunion> reuniones = Reunion.getReuniones(id);
 
-		// Elimina en la BD todos los archivos de las reuniones
-		for (int i = 0; i < reuniones.size(); i++) {
-			Archivo.eliminaTodo(reuniones.get(i).id);
+				// Elimina en la BD todos los archivos de las reuniones
+				for (int i = 0; i < reuniones.size(); i++) {
+					Archivo.eliminaTodo(reuniones.get(i).id);
+				}
+
+				// Se crea un array con todos los archivos dentro del directorio
+				// del grupo.
+				File[] ficheros = file.listFiles();
+
+				// Elimina cada archivo dentro del directorio.
+				for (int x = 0; x < ficheros.length; x++) {
+					ficheros[x].delete();
+				}
+				// Elimina el directorio una vez vacio.
+				file.delete();
+
+				Reunion.eliminaTodo(id);
+				Integrante.eliminaTodos(id);
+				Grupo.find.ref(id).delete();
+				return redirect(routes.Home.index());
+			}
 		}
-
-		// Se crea un array con todos los archivos dentro del directorio
-		// del grupo.
-		File[] ficheros = file.listFiles();
-
-		// Elimina cada archivo dentro del directorio.
-		for (int x = 0; x < ficheros.length; x++) {
-			ficheros[x].delete();
-		}
-		// Elimina el directorio una vez vacio.
-		file.delete();
-
-		Reunion.eliminaTodo(id);
-		Integrante.eliminaTodos(id);
-		Grupo.find.ref(id).delete();
-		return redirect(routes.Home.index());
 	}
 
 	/**
